* Shut some annoying warning.

+++ source/src/SPU.cpp
@@ -621,7 +621,7 @@ template<SPUInterpolationMode INTERPOLATE_MODE> static FORCEINLINE void FetchADP
 	    	chan->pcm16b_last = chan->pcm16b;
 	    	chan->pcm16b = MinMax(chan->pcm16b+diff, -0x8000, 0x7FFF);
 
-			if(i == (chan->loopstart<<3)) {
+			if(i == (u32)(chan->loopstart<<3)) {
 				if(chan->loop_index != K_ADPCM_LOOPING_RECOVERY_INDEX) printf("over-snagging\n");
 				chan->loop_pcm16b = chan->pcm16b;
 				chan->loop_index = chan->index;
